Organize some testing classes into packages
