+++ app/src/androidTest/java/net/osmtracker/data/Mocks.java
@@ -1,0 +1,14 @@
+package net.osmtracker.data;
+
+public class Mocks {
+    public static String MOCK_LAYOUT_CONTENT =
+            "<layouts>"
+                    +"    <layout name=\"root\">"
+                    +"		<row>"
+                    +"			<button type=\"tag\" label=\"A\"/>"
+                    +"			<button type=\"tag\" label=\"B\"/>"
+                    +"			<button type=\"tag\" label=\"C\"/>"
+                    +"		</row>"
+                    +"	</layout>"
+                    +"</layouts>";
+}
