+++ app/src/androidTest/java/net/osmtracker/layouts/DeleteLayoutTest.java
@@ -20,11 +20,11 @@ import static android.support.test.espresso.action.ViewActions.longClick;
 import static android.support.test.espresso.assertion.ViewAssertions.doesNotExist;
 import static android.support.test.espresso.matcher.PreferenceMatchers.withTitleText;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static net.osmtracker.layouts.TestUtils.checkToastIsShownWith;
-import static net.osmtracker.layouts.TestUtils.getLayoutsDirectory;
-import static net.osmtracker.layouts.TestUtils.getStringResource;
-import static net.osmtracker.layouts.TestUtils.injectMockLayout;
-import static net.osmtracker.layouts.TestUtils.listFiles;
+import static net.osmtracker.util.TestUtils.checkToastIsShownWith;
+import static net.osmtracker.util.TestUtils.getLayoutsDirectory;
+import static net.osmtracker.util.TestUtils.getStringResource;
+import static net.osmtracker.util.TestUtils.injectMockLayout;
+import static net.osmtracker.util.TestUtils.listFiles;
 import static org.hamcrest.Matchers.equalToIgnoringCase;
 import static org.junit.Assert.assertFalse;
 import static org.apache.commons.io.FileUtils.deleteDirectory;
