--- app/src/androidTest/java/net/osmtracker/layouts/MockData.java
@@ -1,14 +1,0 @@
-package net.osmtracker.layouts;
-
-public class MockData {
-    public static String MOCK_LAYOUT_CONTENT =
-            "<layouts>"
-            +"    <layout name=\"root\">"
-            +"		<row>"
-            +"			<button type=\"tag\" label=\"A\"/>"
-            +"			<button type=\"tag\" label=\"B\"/>"
-            +"			<button type=\"tag\" label=\"C\"/>"
-            +"		</row>"
-            +"	</layout>"
-            +"</layouts>";
-}
